x86/resctrl: Move resctrl types to a separate header
To avoid sticky problems in the mpam glue code, move the resctrl
enums into a separate header.

This lets the arch code declare prototypes that use these enums without
creating a loop via asm<->linux resctrl.h

The same logic applies to the monitor-configuration defines, move these
too.

(cherry picked from commit 93910dab28b19551cb5b8c2064ea23d7a1a0ff2b https://git.kernel.org/pub/scm/linux/kernel/git/morse/linux.git/log/?h=mpam/snapshot/v6.7-rc2)
